Only write changed colors to the ESource (#19)
When a calendar is first created, the ESource visible to the backend
has the Address Book extension enabled for some reason. This means
that an invalid address book is created. But this only happens when
we also write the ESource to disk. If we minimize doing that, the
bug will occur less often.